+++ src/UploadController.php
@@ -45,7 +45,9 @@ class UploadController extends Controller {
         $image->Filename = $postVars['original_filename'];
         $image->ThumbnailURL = $postVars['thumbnail'];
         $image->Size = $postVars['bytes'];
-        
+        $image->Width = $postVars['width'];
+        $image->Height = $postVars['height'];
+
         $this->extend('onBeforeImageCreated', $image, $postVars);
 
         try {